all: set base dir based on the first .ssh directory that we found
Previously, the base directory is set on current working directory.
This limit the way to execute awwan only from the directory
that contains the .ssh directory.

This changes set the environment BaseDir based on the first .ssh
directory that we found from script path up to the "/".  So, user can
execute awwan script from any sub-directories.