all: use native SFTP to put and get file when possible
Previously, the command to put and get file from remote server depends
on installed scp program on the host computer.

In this changes we add the SFTP client and use it as primary function
to put and get file to/from remote when possible and use the scp as
fallback.