all: refactoring the code
Previously we have Command, environment, and script types to handle
all functionalities.  The name for Command is kind a misleading, because
it contains method to copy files in local or remote and to execute
script in local or remote.

This refactoring break down the types into Awwan, Session, and Script.
Awwan contains cache of sessions and cache of environment files.
Session manage and cache SSH client and list of scripts.
One session have one SSH client, but may contains more than one script.
Script define the content of ".aww" file, line by line.