all: add sub-command "build" to awwan
Previously, we depends external commands to watch, build, and embed the
files in _www directory for development.

This changes simplify all of this into single command.

To compile the TypeScript files and embed them we just run,

  go run ./cmd/awwan build

To run the development server that watch any changes on .ts/.js/.html
files inside _www and rebuild them

  AWWAN_DEVELOPMENT=1 go run ./cmd/awwan serve ${AWWAN_WORKSPACE}