cmd/awwan: make the "help" and "version" as command
Previously, to print the help and version we use the flag package, "-h"
for printing usage and "-v" for printing "version".

The flag should be used to changes the behaviour of command not
terminating the command.
Using and mixing flag as command make the CLI API to be inconsistent.

For example, the following argument is inconsistent,

  $ awwan -h play script.aww 1 -

Instead of executing script.aww, it print the usage and terminate.