all: make .Vars, .Val, and .Vals panic if values is empty
This is to prevent copying or executing command with value that are
not defined or typo which make the result empty and may result in
undefined behaviour.

For example if we have "app_dir = /data/app" and command in the script
that remove that directory recursively,

  sudo rm -r {{.Val "::app_dir}}/bin

will result removing "/bin" entirely.