all: simplify parsing statement for magic command get
This changes minimize code duplication when parsing magic command
between "#get:" and "#get!" using single function parseStatementGetPut.

In the returned Statement, we changes where to store the source argument
into args[0] instead of in cmd for better readability and minimize
confusion in Copy/SudoCopy.

While at it, we add unit test for "#get:".
Unit test for "#get!" is not possible right now, because it will prompt
for password.