session: use the raw statement when adding "-S" option in ExecLocal
Previously, when refactoring ExecLocal in bf9c8226b4d to support sudo
with -S option, we generate the raw command by concatenating Statement
args field.
This model does not works if the raw command contains double quote that
needs to be executed as is.

This changes fix this issue by adding "-S" option to raw command.