_mkosi: populate SSH keys manually using "mkosi.extra"
Previously, when building container, the SSH keys for server and user
are generated automatically.
This makes the SSH keys unpredictable when we need to test SSH function
later (the known_hosts for source and authorized_keys in target).

This changes set the host and user SSH keys manually using "mkosi.extra"
skeleton, so every container build, will have predictable keys.