all: return the error as reponse in HTTP API execute
Previously, when the command execution failed, we check the error and
return it as HTTP status code 500.
In this way, user cannot view the log and actual error.

In this changes, if the command failed, we store the error in separate
field "Error" and return to the caller with HTTP status code 200.