all: refactoring Local and Play so that error get logged
Previously, if there is any error from Local or Play before any statement
get executed it will not logged to log, also the Request close() is
not called since we put it as defer in the middle.

This changes fix this by moving down error handler into single exit
goto label.