internal: use separate DirWatcher instead of using instance of MemfsWui
In MemfsWui we exclude the "adoc" files but when calling Watch
we pass it as to be included.
When the adoc file changes, sometimes the changes not detected, probably
there is a bug in the memfs.Watch.
For now, we use separate instance of DirWatcher to watch any changes
on adoc, html, js, md, and files.