all: always load SSH config when running Play
In case awwan run with "serve" and we modify the ".ssh/config", the
changes does not detected by awwan because we only read ".ssh/config"
once we Awwan instance created.

This changes fix this issue by always loading SSH config everytime
the Play method executed so the user CLI and WUI has the same experiences.