_ops: setup awwan-play
The _ops/awwan-play is the seed for the awwan playground that will
be live on "play.awwan.org" later.

The image name is awwan-play, to allow us to serve play.awwan.local
on local proxy and test it without port number.

It use the _play directory as workspace.

References: https://todo.sr.ht/~shulhan/awwan/4