all: refactoring container for integration tests
Instead of building the container multiple times, create it one time only
that importable to the system /var/lib/machines.
So future development can enable or start them manually and run the
test-integration or test-all tasks without rebuilding them again.

This changes require moving the image output to /data/awwan because if
the test image is located under the sub directory of the source
directory and we bind the source directory into container it will cause
recursive mount that makes its impossible to use the image inside
/var/lib/machines.

We also set the network zone to "awwan", to allow multiple containers
shared the same subnet on the host later.