_wui: use the output for displaying notification
Previously, we use a quick "pop-up" to display notification for each
information or error from WUI.
Sometimes this is annoying, it overlap the buttons, make it hard to
Save and Encrypt at the same time.

In this changes we move the nofication message to be displayed in
the output, same with output of execution.