all: reduce the response on HTTP endpoint on GET fs
Previously, the HTTP endpoint for "GET /awwan/api/fs" return the content
of files when the requested node is a directory.
This is cause unnecessary load because when requesting directory we
only need list of file names not the content.

This changes reduce the response by returning only list of node child
without its content, which require update on share module on [Node.JSON].