all: make the magic line "#local" works on "local" command too
In case we have a script that manage local host and remote server,
calling "play" on "#local" lines only always open the connection to remote
server.

To minimize opening unused connections, let the "#local" command works
on both commands.  Its up to user which part of lines that they want
to execute on remote or local.