all: make the magic line "#put:" use explicit source for encrypted file
Previously, the magic line "#put" detect whether the source file is
encrypted or not automatically, so we did not need to put ".vault"
suffix in the source path.
This changes make it to be always explicit.
If we want to copy encrypted file than the source file must be the path
to encrypted file.

Implements: https://todo.sr.ht/~shulhan/awwan/7