all: change the remote temporary directory to "~/.cache/awwan"
If the file to be copied contains sensitive data, putting them in
"/tmp" considered a security risk, even though it will be moved to
destination later.

The issue is when the "#put" command failed, the plain file is left
on "/tmp" directory.

This changes add additional advantage where we did not need to remove
the temporary directory on remote when execution completed, since the
temporary directory should be accessible by user only.

Implements: https://todo.sr.ht/~shulhan/awwan/8