all: update machine container setup with latest Arch
The latest mkosi use importctl instead of machinectl to import the
image.

There is an issue in mkosi.conf where build with Format=tar always get
compressed using zstd, but importctl unable to import it, so we need
to set CompressOutput=false.

The home directory mode needs to be set to open 0755, because now its
default to 0700, which makes the test failed to read file on others
user's home.

The pacman mirror changes to public one that is always available,
geo.mirror.pkgbuild.com.

The BUG_REPORT_URL in os-release for Arch changes which we need to
modify because its included in the test-container.